requirements.txt
bcrypt==4.0.1
boto3==1.26.50
botocore==1.29.50
click==8.1.3
cmake==3.25.0
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
numpy==1.24.1
Pillow==9.4.0
protobuf==3.20.1
pyotp==2.8.0
python-dateutil==2.8.2
python-decouple==3.7
qrcode==7.3.1
s3transfer==0.6.0
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
SQLAlchemy==1.4.46
threadpoolctl==3.1.0
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.11.0
